Completado middleware de validaciones y verificacion de rol, creada interface para extender JwtPyload, movido HttpResponse a carpeta shared, modulo de notFound 404 siempre al final del array<Router>

+++ src/modules/Users/users_controller.ts
@@ -1,7 +1,6 @@
-import { type Request, type Response } from 'express';
-import { type HttpResponse } from '../HttpResponse';
-import { UserService } from './users_service';
-import { loginType } from './users_schemas';
+import type { Request, Response } from 'express';
+import type { HttpResponse } from '../shared/httpResponse';
+import type { UserService } from './users_service';
 
 export class UsersController {
   readonly #response: HttpResponse;
@@ -13,12 +12,13 @@ export class UsersController {
 
   auth = async (req: Request, res: Response): Promise<Response> => {
     try {
-      const { username, password }: loginType = req.body;
+      const { username, password } = req.body;
       if (username === undefined || password === undefined) return this.#response.badRequest(res);
       const user = await this.#service.checkUserAndPassword(username, password);
       if (user === undefined) return this.#response.unauthorized(res, 'Username or password is incorrect');
+      const { _id, email, name, roles, telf } = user;
       const token = this.#service.generateJwt(user);
-      return this.#response.ok(res, { token });
+      return this.#response.ok(res, { token, _id, email, username, name, roles, telf });
     } catch (error) {
       console.error(error);
       return this.#response.error(res, error);
@@ -27,11 +27,9 @@ export class UsersController {
 
   getOneUser = async (req: Request, res: Response): Promise<Response> => {
     const { _id } = req.params;
-    //console.log(_id);
     try {
       const user = await this.#service.getUserById(_id);
       if (user === undefined) return this.#response.notFound(res);
-      console.log(user);
       return this.#response.ok(res, user);
     } catch (error) {
       return this.#response.error(res, error);
@@ -41,7 +39,6 @@ export class UsersController {
   getUsers = async (_req: Request, res: Response): Promise<Response> => {
     try {
       const users = await this.#service.getAllUsers();
-      console.log(users);
       return this.#response.ok(res, users);
     } catch (error) {
       return this.#response.error(res, error);
@@ -60,9 +57,8 @@ export class UsersController {
 
   updateUser = async (req: Request, res: Response): Promise<Response> => {
     const { _id } = req.params;
-    const { body } = req;
     try {
-      const user = await this.#service.updateUserById(_id, body);
+      const user = await this.#service.updateUserById(_id, req.body);
       if (user === undefined) return this.#response.notFound(res);
       return this.#response.ok(res, user);
     } catch (error) {
@@ -72,10 +68,10 @@ export class UsersController {
 
   deleteUser = async (req: Request, res: Response): Promise<Response> => {
     const { _id } = req.params;
+    if (_id.length < 24) return this.#response.badRequest(res, 'Argument passed in must be a 24 hex characters');
     try {
       const user = await this.#service.deleteUserById(_id);
       if (user === undefined) return this.#response.notFound(res);
-      console.log(user);
       return this.#response.ok(res, 'Deleted');
     } catch (error) {
       return this.#response.error(res, error);
